Remove asset handling, JSX transform and Hydration responsibilities. (also remove Generators)
Asset Handling (including the React JS) should be the application’s responsibility.
Doing it from this gem causes version conflicts. Since the rest of the app uses
React from Browserify, overriding it also isn’t an option. It is IMO ideal to move to
NPM+CommonJS for all JS asset handling.

Transforming JSX is also better handled by the application, or as a separate gem.
In my case I let `browserify-rails` do this using the `reactify` transform.

Hydration has to be inside the Browserify bundle since it’ll need to refer
to the React components. To make it work with Browserify, we’ll have to push
all the React components required for server-side rendering into a global namespace.
That will cause name conflicts in large projects, and will nullify any benefits
of Browserify. Also I prefer to let the application control hydration instead of
hiding it behind a gem. The app can easily bundle hydration along with
the rest of the Browserify server bundle to give it access to all required
components.

I also removed Generators so that this gem will be responsible only for
pre-rendering of React components.